WED-9 Refactor code structure and Readme
