+++ .env.local.example
@@ -3,3 +3,6 @@ CLOUDINARY_API_KEY=
 CLOUDINARY_API_SECRET=
 WEDDING_FOLDER=
 SLIDER_FOLDER=
+AIRTABLE_API_KEY=
+AIRTABLE_BASE_ID=
+AIRTABLE_TABLE_NAME=
