--- components/Footer.tsx
@@ -1,9 +1,0 @@
-const Footer = () => {
-  return (
-    <footer className="p-4 md:px-6 md:py-8">
-      <span className="block text-sm text-gray-500 text-center"><i>#vuandtrang 2023</i></span>
-    </footer>
-  );
-};
-
-export default Footer;
