+++ components/LandingPage/Introduction.tsx
@@ -1,3 +1,5 @@
+import { introduction } from "../../utils/config";
+
 const Introduction = () => {
   return (
     <section className="py-8 px-10 md:px-48 text-gray-700">
@@ -5,8 +7,7 @@ const Introduction = () => {
         We are Getting Married!
       </h1>
       <p className="text-center">
-        Chúng mình rất hạnh phúc khi lên từng chi tiết cho lễ thành hôn với người mà mình yêu nhất, trước sự chứng kiến của gia đình và những người bạn thân thiết.
-        Cám ơn các bạn đã dành thời gian ghé thăm trang web của chúng mình!
+        {introduction}
       </p>
     </section>
   );
