+++ pages/_document.tsx
@@ -1,4 +1,5 @@
-import Document, { Head, Html, Main, NextScript } from 'next/document'
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+import { metadata } from '../utils/config';
 
 class MyDocument extends Document {
   render() {
@@ -8,16 +9,11 @@ class MyDocument extends Document {
           <link rel="icon" href="/logo.svg" />
           <link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=Great Vibes&display=swap" />
           <link rel="stylesheet" type="text/css" href="https://fonts.googleapis.com/css?family=Muli&display=swap" />
-          <meta
-            name="description"
-            content="VuandTrang Wedding Site!"
-          />
-          <meta property="og:site_name" content="vandt.day" />
-          <meta
-            property="og:description"
-            content="VuandTrang Wedding Site!"
-          />
-          <meta property="og:title" content="VuandTrang Wedding Site!" />
+          <meta name="description" content={metadata.description} />
+          <meta property="og:site_name" content={metadata.ogSitename} />
+          <meta property="og:description" content={metadata.description} />
+          <meta property="og:title" content={metadata.title} />
+          <meta property="og:image" content={metadata.ogImage} />
         </Head>
         <body className="antialiased">
           <Main />
@@ -28,4 +24,4 @@ class MyDocument extends Document {
   }
 }
 
-export default MyDocument
+export default MyDocument;
