--- utils/common.ts
@@ -1,43 +1,0 @@
-import groom from '../public/groom.webp';
-import bride from '../public/bride.webp';
-
-export const groomAndBride = [
-  {
-    image: bride,
-    alt: 'Cô dâu',
-    name: 'Vũ Minh Trang'
-  },
-  {
-    image: groom,
-    alt: 'Chú rể',
-    name: 'Đoàn Thế Vũ'
-  }
-];
-
-export const ourStories = [
-  {
-    time: '3/12/2017',
-    title: 'Lần đầu gặp gỡ',
-    description: `Một lần rất tình cờ chúng mình được những người bạn chung rủ đi chơi theo tiêu chí “càng đông càng vui”. Thế là thành lần gặp gỡ đầu tiên, tuy không dài nhưng đủ để cho 2 đứa có ấn tượng tốt về “bạn kia” 😊 \n
-Và ấn tượng tốt đấy đưa bọn mình thành bạn bè, tới những cuộc hội thoại vu vơ mỗi buổi tối.`
-  },
-  {
-    time: '2/4/2020',
-    title: 'Chính thức hẹn hò',
-    description: `Dần dần có một cảm giác trên tình bạn đã dần len lỏi vào từng câu chuyện nhỏ của chúng mình. \n
-Nhưng ở thời điểm bắt đầu bước chân vào “thị trường lao động”, cả 2 đều muốn tập trung vào sự nghiệp. Những cuộc trò chuyện cũng không còn liên tục, nhưng 2 đứa vẫn luôn quan tâm tới cuộc sống của đối phương. \n
-Và rồi, một ngày mang ý nghĩa là hậu cá tháng 4, bọn mình đã chính thức đến với nhau 😊`
-  },
-  {
-    time: '14/10/2022',
-    title: 'Cầu hôn',
-    description: `Suốt thời gian gần 3 năm, 2 đứa bên nhau cùng đi qua những cung bậc cảm xúc của tình yêu, cũng như là người bạn đồng hành qua những trải nghiệm trong công việc và cuộc sống. Bọn mình đều cùng nhau đi qua, tự hoàn thiện bản thân để trở thành “better me” và thành chỗ dựa cho người còn lại. \n
-Ngày 14/10, đằng gái chính thức nhận lời cầu hôn, cả 2 đã sẵn sàng xây dựng hạnh phúc với nửa kia.`
-  },
-  {
-    time: '11/3/2023',
-    title: 'Lễ thành hôn',
-    description: `Không có một từ ngữ nào có thể diễn tả cảm xúc của chúng mình trong khoảng thời gian chờ đợi cho đến ngày trọng đại nhất của cuộc đời, cùng nhau viết nên một chặng đường mới với người mà mình yêu. \n
-Dù cho có khó khăn như thế nào, cũng sẽ có một người cùng mình nắm tay đi về phía trước.`
-  }
-]
